refactoring to simplify structure updated to use latest parent bom (springboot 3)
