--- model/src/main/java/model.uml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Diagram>
-  <ID>JAVA</ID>
-  <OriginalElement>D:/_programming/demo-rest2/model/src/main/java/ca/mikegabelmann/demo2/persistence/model</OriginalElement>
-  <nodes>
-    <node x="0.0" y="126.0">ca.mikegabelmann.demo2.persistence.model.Person</node>
-    <node x="224.0" y="0.0">ca.mikegabelmann.demo2.persistence.model.GroupTypeCodeId</node>
-    <node x="12.5" y="362.0">ca.mikegabelmann.demo2.persistence.model.Address</node>
-    <node x="417.0" y="23.5">ca.mikegabelmann.demo2.persistence.model.BooleanConverter</node>
-    <node x="606.0" y="0.0">ca.mikegabelmann.demo2.persistence.model.GroupCode</node>
-    <node x="368.5625" y="136.0">ca.mikegabelmann.demo2.persistence.model.GroupTypeCode</node>
-    <node x="18.5" y="0.0">ca.mikegabelmann.demo2.persistence.model.SexCode</node>
-  </nodes>
-  <notes />
-  <edges>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.Person" target="ca.mikegabelmann.demo2.persistence.model.SexCode" relationship="TO_ONE">
-      <point x="15.0" y="-93.0" />
-      <point x="15.0" y="38.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.GroupTypeCode" target="ca.mikegabelmann.demo2.persistence.model.BooleanConverter" relationship="TO_ONE">
-      <point x="1.9375" y="-93.0" />
-      <point x="0.0" y="14.5" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.Person" target="ca.mikegabelmann.demo2.persistence.model.Address" relationship="TO_MANY">
-      <point x="-15.0" y="93.0" />
-      <point x="-15.0" y="-104.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.GroupTypeCode" target="ca.mikegabelmann.demo2.persistence.model.GroupTypeCodeId" relationship="DEPENDENCY">
-      <point x="-108.5" y="-93.0" />
-      <point x="80.5625" y="38.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.GroupTypeCode" target="ca.mikegabelmann.demo2.persistence.model.BooleanConverter" relationship="DEPENDENCY">
-      <point x="16.9375" y="-93.0" />
-      <point x="15.0" y="14.5" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.Person" target="ca.mikegabelmann.demo2.persistence.model.SexCode" relationship="DEPENDENCY">
-      <point x="0.0" y="-93.0" />
-      <point x="0.0" y="38.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.GroupTypeCode" target="ca.mikegabelmann.demo2.persistence.model.GroupCode" relationship="DEPENDENCY">
-      <point x="123.5" y="-93.0" />
-      <point x="-53.9375" y="38.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.Person" target="ca.mikegabelmann.demo2.persistence.model.Address" relationship="DEPENDENCY">
-      <point x="30.0" y="93.0" />
-      <point x="30.0" y="-104.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.Address" target="ca.mikegabelmann.demo2.persistence.model.Person" relationship="DEPENDENCY">
-      <point x="15.0" y="-104.0" />
-      <point x="15.0" y="93.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.GroupTypeCode" target="ca.mikegabelmann.demo2.persistence.model.GroupCode" relationship="TO_ONE">
-      <point x="108.5" y="-93.0" />
-      <point x="-68.9375" y="38.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.Address" target="ca.mikegabelmann.demo2.persistence.model.Person" relationship="TO_ONE">
-      <point x="0.0" y="-104.0" />
-      <point x="0.0" y="93.0" />
-    </edge>
-    <edge source="ca.mikegabelmann.demo2.persistence.model.GroupTypeCode" target="ca.mikegabelmann.demo2.persistence.model.GroupTypeCodeId" relationship="TO_ONE">
-      <point x="-123.5" y="-93.0" />
-      <point x="65.5625" y="38.0" />
-    </edge>
-  </edges>
-  <settings layout="Hierarchic" zoom="1.0" showDependencies="true" x="374.0" y="285.0" />
-  <SelectedNodes />
-  <Categories>
-    <Category>Properties</Category>
-  </Categories>
-  <SCOPE>All</SCOPE>
-  <VISIBILITY>private</VISIBILITY>
-</Diagram>
-
