+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/facade/PersonFacade.java
@@ -1,0 +1,17 @@
+package ca.mikegabelmann.demo2.persistence.facade;
+
+import ca.mikegabelmann.demo2.persistence.service.dto.PersonAddress;
+
+import java.util.Optional;
+
+
+public interface PersonFacade {
+
+    /**
+     * Get person and addresses given person id.
+     * @param id person id
+     * @return record if it exists
+     */
+    Optional<PersonAddress> getPersonAddress(Long id);
+
+}
