+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/service/AddressService.java
@@ -1,0 +1,12 @@
+package ca.mikegabelmann.demo2.persistence.service;
+
+import ca.mikegabelmann.demo2.persistence.model.Address;
+
+import java.util.List;
+
+
+public interface AddressService {
+
+    List<Address> getAddressByCountryAndProvAndCity(String country, String prov, String city);
+
+}
