+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/service/PersonService.java
@@ -1,0 +1,20 @@
+package ca.mikegabelmann.demo2.persistence.service;
+
+import ca.mikegabelmann.demo2.persistence.model.Person;
+
+import java.time.LocalDate;
+import java.util.List;
+
+
+public interface PersonService {
+
+    /**
+     * Find people by sex and birth date.
+     * @param sex sex
+     * @param date birth date
+     * @return records
+     */
+    List<Person> findBySexAndBirthDt(String sex, LocalDate date);
+
+    Person createOrUpdate(Person p);
+}
