+++ persistence/src/test/java/ca/mikegabelmann/demo2/persistence/repository/GroupCodeRepositoryTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import javax.persistence.EntityNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 
 @DataJpaTest
