+++ pom.xml
@@ -30,13 +30,11 @@
     </properties>
 
     <modules>
-        <module>model</module>
         <module>shared</module>
         <module>persistence</module>
-        <module>service</module>
         <module>rest</module>
-        <!--<module>client</module>
-        <module>app</module>-->
+        <!--<module>client</module>-->
+        <!--<module>app</module>-->
     </modules>
 
 </project>
