+++ rest/src/main/java/ca/mikegabelmann/demo2/controller/config/CorsConfig.java
@@ -1,6 +1,7 @@
 package ca.mikegabelmann.demo2.controller.config;
 
 import ca.mikegabelmann.demo2.controller.rest.HelloWorldRestController;
+import jakarta.annotation.PostConstruct;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,8 +14,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import javax.annotation.PostConstruct;
 
 
 /**
