+++ rest/src/main/java/ca/mikegabelmann/demo2/controller/config/OtherEnvironmentConfig.java
@@ -1,12 +1,11 @@
 package ca.mikegabelmann.demo2.controller.config;
 
+import jakarta.annotation.PostConstruct;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-
-import javax.annotation.PostConstruct;
 
 
 @Profile("!local")
