+++ rest/src/test/java/ca/mikegabelmann/demo2/controller/rest/mapper/DtoMapperTest.java
@@ -6,7 +6,7 @@ import ca.mikegabelmann.demo2.dto.PersonDto;
 import ca.mikegabelmann.demo2.persistence.model.Address;
 import ca.mikegabelmann.demo2.persistence.model.Person;
 import ca.mikegabelmann.demo2.persistence.model.SexCode;
-import ca.mikegabelmann.demo2.service.dto.PersonAddress;
+import ca.mikegabelmann.demo2.persistence.service.dto.PersonAddress;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
