--- service/src/main/java/ca/mikegabelmann/demo2/service/AddressService.java
@@ -1,12 +1,0 @@
-package ca.mikegabelmann.demo2.service;
-
-import ca.mikegabelmann.demo2.persistence.model.Address;
-
-import java.util.List;
-
-
-public interface AddressService {
-
-    List<Address> getAddressByCountryAndProvAndCity(String country, String prov, String city);
-
-}
